`modify-geometry` → `move-vertex`
