+++ components/FormRelation.vue
@@ -275,7 +275,7 @@
   import { OpenFormStep }                                 from '../actions/open-form';
   import { OpenTableStep }                                from '../actions/open-table';
   import { AddFeatureStep }                               from '../actions/add-feature';
-  import { ModifyGeometryVertexStep }                     from '../actions/modify-geometry';
+  import { ModifyGeometryVertexStep }                     from '../actions/move-vertex';
   import { MoveFeatureStep }                              from '../actions/move-feature';
 
   const { ProjectsRegistry }            = g3wsdk.core.project;
